Further improvements to scraper to more comprehensively grab cards
Previously, scraper was topping out at roughly 10k cards regardless of page number available. This is due to TCGPlayer most likely limiting the number of pages deep you can go based on total cards in that filter. So, with 24 per page, it was 416 pages, 36 it was 277 pages, and with 48 it was around 208. We can get around this, hypothetically, by going by set. Currently running a test against 20k+ card store to assess.